requirements.txt
streamlit
winrt-Windows.Media.Control
winrt-Windows.Foundation
groq
ytmusicapi
python-dotenv
plotly
requests
